+++ metadata.rb
@@ -12,4 +12,4 @@ supports 'ubuntu', '>= 18.04'
 supports 'debian', '>= 9.0'
 supports 'centos', '>= 7.0'
 
-depends 'selinux_policy', '~> 2.4'
+depends 'selinux', '>= 6.0'