+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+- Testing MariaDB 10.4 server install
 - Fix root password cannot be manually set on 10.4+
 
 ## 5.1.0 - *2021-10-13*